feat(rp): add WithPKCEFromDisocvery
Add the WithPKCEFromDiscovery option to create a relying party with
PKCE enabled if it is supported when query the discovery endpoint as
discussed in #506.

This only works when creating an OIDC RP which performs a discovery
call. With an OAuth2-only RP, an error is returned as no discovery call
is performed.

Closes #506